+++ test/numberto/printers_test.clj
@@ -13,12 +13,14 @@
   (is (thrown? IllegalArgumentException (p/format-num [])))
   (is (thrown? IllegalArgumentException (p/format-num -1))))
 
-(deftest number-name-test
-  (is (= "" (p/number-name 0)))
-  (is (= "one" (p/number-name 1)))
-  (is (= "seventeen" (p/number-name 17)))
-  (is (= "one million seventeen" (p/number-name 1000017)))
-  (is (= "fifty four thousand three hundred twenty one" 
-         (p/number-name 54321)))
-  (is (= "one duotrigintillion" 
-         (p/number-name (reduce *' (repeat 99 10))))))
+(deftest format-ratio-test
+  (is (= "0.3" (p/format-ratio 1/3 1)))
+  (is (= "0.33" (p/format-ratio 1/3 2)))
+  (is (= "0.67" (p/format-ratio 2/3 2)))
+  (is (= "20.00" (p/format-ratio 100/5 2)))
+  (is (= "1.00" (p/format-ratio 1 2)))
+  (is (= "1.00" (p/format-ratio (- 1 1/10000000) 2)))
+  (is (= "0.99" (p/format-ratio (- 1 1/100) 2)))
+  (is (= "1" (p/format-ratio (- 1 1/10000000) 0)))
+  (is (= "1.000000" (p/format-ratio (- 1 1/10000000) 6)))
+  (is (thrown? IllegalArgumentException (p/format-ratio 0.5 2))))